fix: multi-notebook replay does not properly handle tab titles
due to the way replay loads notebooks, versus commentary; this PR consolidates this logic
