+++ plugins/plugin-client-common/.gitignore
@@ -1,1 +1,2 @@
 dist
+mdist
