+++ plugins/plugin-client-common/notebook/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@kui-shell/plugin-client-common-notebook",
+  "version": "6.0.12",
+  "description": "Kui plugin that offers notebook utilities",
+  "license": "Apache-2.0",
+  "author": "Nick Mitchell",
+  "homepage": "https://github.com/IBM/kui#readme",
+  "bugs": {
+    "url": "https://github.com/IBM/kui/issues/new"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/IBM/kui.git"
+  },
+  "keywords": [
+    "kui",
+    "plugin"
+  ],
+  "contributors": [
+    "Mengting Yan"
+  ],
+  "main": "dist/index.js",
+  "module": "mdist/index.js",
+  "types": "mdist/index.d.ts",
+  "publishConfig": {
+    "access": "private"
+  }
+}
