+++ plugins/plugin-client-common/notebook/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../../packages/builder/tsconfig-base.json",
+  "include": ["src/**/*"],
+  "compilerOptions": {
+    "composite": true,
+    "outDir": "mdist",
+    "rootDir": "src"
+  }
+}
